+++ scripts/lib/_menu.psm1
@@ -16,4 +16,4 @@ $ModuleDir = $PSScriptRoot
 . (Join-Path $ModuleDir 'menu\status-view.ps1')
 . (Join-Path $ModuleDir 'menu\text-input.ps1')
 
-Export-ModuleMember -Function Start-MenuSession, Stop-MenuSession, Suspend-MenuSession, Select-Worktree, Select-Action, Select-YesNo, Select-Grid, Show-GridStatus, Show-MenuStatus, Read-MenuText
+Export-ModuleMember -Function Start-MenuSession, Stop-MenuSession, Suspend-MenuSession, Select-Worktree, Select-Action, Select-YesNo, Select-Grid, New-GridStatusView, Update-GridStatusView, Show-GridStatus, Show-MenuStatus, Read-MenuText